Allow a delete to proceed if file is locked by removing the associated xattr metadata item.
In the case a file has an existing (stuck) lock, when the file is requested to be deleted, it will fail with -EBUSY. In this case we remove the lock and retry the unlink command.
This asserts therefore that a delete should take precedence over other activities.
As libradosstriper uses shared locks for both reads and writes, there is little chance to catch the few edge cases where this behaviour may be less desirable.
The lock removal is achieved through the removal of the lock extended attribute, using libradosstriper methods.